plugins/orchestration-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.3
httpx>=0.24.0,<0.25.0
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.5.0
pydantic-settings==2.1.0
apscheduler==3.10.4
websockets==12.0
python-socketio==5.10.0
python-multipart==0.0.6
aiohttp==3.9.1
